backend/requirements/base.txt
Django==5.1.15
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.15.0
django-environ==0.13.0
asgiref==3.11.1
PyJWT==2.12.1
sqlparse==0.5.5
python-dateutil==2.9.0.post0
